JUDDI-520, adding a confirmation box on delete
git-svn-id: https://svn.apache.org/repos/asf/juddi/trunk@1167018 13f79535-47bb-0310-9956-ffa450edef68

+++ juddi-console/uddi-portlets/src/main/java/org/apache/juddi/portlets/client/PublisherPanel.java
@@ -135,20 +135,22 @@ public class PublisherPanel extends FlowPanel {
 	
 	protected void deletePublisher(String token){
 		if (publisher!=null) {
-			juddiApiService.deletePublisher(token, publisher.getAuthorizedName(), new AsyncCallback<JUDDIApiResponse>() 
-			{
-				public void onFailure(Throwable caught) {
-					Window.alert("Error: " + caught.getMessage());
-				}
-	
-				public void onSuccess(JUDDIApiResponse response) {
-					if (response.isSuccess()) {
-						JUDDIPublisher.getInstance().hidePublisher();
-					} else {
-						Window.alert("error: " + response.getMessage());
+			if (Window.confirm("Are you sure you want to delete Publisher '" + publisher.getAuthorizedName() + "'?")) {
+				juddiApiService.deletePublisher(token, publisher.getAuthorizedName(), new AsyncCallback<JUDDIApiResponse>() 
+				{
+					public void onFailure(Throwable caught) {
+						Window.alert("Error: " + caught.getMessage());
 					}
-				}
-			});
+		
+					public void onSuccess(JUDDIApiResponse response) {
+						if (response.isSuccess()) {
+							JUDDIPublisher.getInstance().hidePublisher();
+						} else {
+							Window.alert("error: " + response.getMessage());
+						}
+					}
+				});
+			}
 		}
 	}
